PlotterMonitor: only autoRange if no max value was given
Currently, the PlotterMonitor does auto-ranging of the plotter. However,
if we explicitly specified a maximum value, that doesn't make much
sense, because it renders the maximum value pretty much ineffective.
Therefore, enable autoRange only if max_value is not set.

This requires us to set an explicit rangeMin and rangeMax. rangeMin is
forced to zero (which may or may not be appropriate, but we currently
don't have a configuration option for it). rangeMax is set to 1 because
the data is normalized so that it is 1 if it reaches max_value.

Note that if autoRange is true, rangeMin and rangeMax are ignored.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>